refactor: rename private nplikes
